Add 'pull-forward' command
Implement 'pull-forward' command. This is similar to 'pull --ff-only',
except rather than relying on a tracking branch, it considers the branch
of the same name on all remotes. This is extremely useful in the common
case where the branch is regularly pushed to (and is tracking) a fork,
but the usual work flow is to integrate changes from an upstream
repository.